# src/main/resources/static/openapi.yml
openapi: 3.1.0

info:
  title: AuditLog Service API
  description: APIs for tamper evident audit log service
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: local server to access rest apis

paths:
  /v1/audit/events:
    post:
      operationId: createAuditEvent
      summary: Record an audit event
      description: >-
        Appends a new event to the tamper-evident audit log. The server assigns
        the event id, sequence number, and content hash, and links the event to
        the previous entry via previousHash to form a verifiable hash chain.
      tags:
        - Audit Events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditEventRequest'
            example:
              eventType: RECORD_UPDATED
              actorId: user-42
              resourceType: Customer
              resourceId: cust-1001
              payload:
                field: address
                oldValue: 123 Main St
                newValue: 456 Oak Ave
              timestamp: '2026-09-14T10:15:30Z'
      responses:
        '201':
          description: Audit event recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventResponse'
              example:
                id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                sequenceNumber: 1024
                eventType: RECORD_UPDATED
                actorId: user-42
                resourceType: Customer
                resourceId: cust-1001
                payload:
                  field: address
                  oldValue: 123 Main St
                  newValue: 456 Oak Ave
                timestamp: '2026-09-14T10:15:30Z'
                previousHash: a94a8fe5ccb19ba61c4c0873d391e987982fbbd3a94a8fe5ccb19ba61c4c0873
                contentHash: 9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08
        '400':
          description: Invalid audit event request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      operationId: listAuditEvents
      summary: Query audit events
      description: >-
        Returns a cursor-paginated, chronologically ordered page of audit events.
        Events are ordered by ascending sequence number, which is the append order
        of the tamper-evident log. All filters are optional and are combined with
        AND semantics; omitting every filter returns the whole log, page by page.


        Paging is keyset (cursor) based rather than offset based: pass the
        nextCursor returned by the previous response back as the cursor parameter
        to fetch the next page. Because the log is append-only and sequence
        numbers never change, this guarantees no event is skipped or repeated
        while a client is walking the log, and stays fast at any depth.
      tags:
        - Audit Events
      parameters:
        - name: actorId
          in: query
          required: false
          description: Only return events caused by this actor
          schema:
            type: string
          example: user-42
        - name: resourceType
          in: query
          required: false
          description: Only return events affecting this type of resource
          schema:
            type: string
          example: Customer
        - name: resourceId
          in: query
          required: false
          description: Only return events affecting this specific resource
          schema:
            type: string
          example: cust-1001
        - name: eventType
          in: query
          required: false
          description: Only return events of this type
          schema:
            type: string
          example: RECORD_UPDATED
        - name: from
          in: query
          required: false
          description: Only return events with a timestamp at or after this instant (inclusive)
          schema:
            type: string
            format: date-time
          example: '2026-09-01T00:00:00Z'
        - name: to
          in: query
          required: false
          description: Only return events with a timestamp strictly before this instant (exclusive)
          schema:
            type: string
            format: date-time
          example: '2026-10-01T00:00:00Z'
        - name: cursor
          in: query
          required: false
          description: >-
            Sequence number of the last event of the previous page; only events
            with a strictly greater sequence number are returned. Omit it to read
            from the start of the log, and otherwise pass back the nextCursor
            value from the previous response.
          schema:
            type: integer
            format: int64
            minimum: 0
          example: 1024
        - name: pageSize
          in: query
          required: false
          description: Maximum number of events to return per page
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
          example: 20
      responses:
        '200':
          description: Page of matching audit events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventPage'
              example:
                content:
                  - id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    sequenceNumber: 1024
                    eventType: RECORD_UPDATED
                    actorId: user-42
                    resourceType: Customer
                    resourceId: cust-1001
                    payload:
                      field: address
                      oldValue: 123 Test Dr
                      newValue: 456 Oak Ave
                    timestamp: '2026-09-14T10:15:30Z'
                    previousHash: a94a8fe5ccb19ba61c4c0873d391e987982fbbd3a94a8fe5ccb19ba61c4c0873
                    contentHash: 9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08
                pageSize: 20
                nextCursor: 1024
                hasMore: true
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/audit/verify:
    get:
      operationId: verifyAuditLog
      summary: Verify the integrity of the audit log
      description: >-
        Walks the audit log in sequence order and recomputes each event's
        content hash and previousHash link. Returns whether the chain is intact,
        how many records were checked, and details of the first violation found
        (if any).
      tags:
        - Audit Events
      responses:
        '200':
          description: Verification completed; the result reports whether the log is intact
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditVerificationResponse'
              examples:
                valid:
                  summary: Hash chain is intact
                  value:
                    valid: true
                    recordsChecked: 15042
                    firstViolation: null
                invalid:
                  summary: Tampering detected
                  value:
                    valid: false
                    recordsChecked: 42
                    firstViolation:
                      sequenceNumber: 42
                      recordId: '019...'
                      type: CONTENT_HASH_MISMATCH
                      message: Stored content hash does not match canonical event content.

  /v1/audit/export:
    get:
      operationId: exportAuditEvents
      summary: Export audit events as a self-contained verifiable bundle
      description: >-
        Returns every audit event matching exactly one filter — resourceId or
        actorId — as a versioned bundle that can be verified independently of
        this service. Records are emitted in ascending sequence order, which is
        their position in the global tamper-evident hash chain.


        A filtered export is usually NOT a contiguous segment of the global
        chain, so each record keeps its original previousHash. Where exported
        sequence numbers are consecutive the records are adjacent in the global
        chain and the link current.previousHash == prior.contentHash must hold;
        where they are not, previousHash is an external reference to a record
        the filter excluded. The chain metadata anchors the segment: the hash
        preceding the first exported record (firstPreviousHash) and the content
        hash of the last one (lastRecordHash) let a recipient pin the export to
        the chain without seeing the excluded records.
      tags:
        - Audit Events
      parameters:
        - name: resourceId
          in: query
          required: false
          description: >-
            Export the history of this specific resource. Exactly one of
            resourceId or actorId must be supplied; blank values count as absent.
          schema:
            type: string
          example: cust-1001
        - name: actorId
          in: query
          required: false
          description: >-
            Export everything this actor caused. Exactly one of resourceId or
            actorId must be supplied; blank values count as absent.
          schema:
            type: string
          example: user-42
      responses:
        '200':
          description: >-
            The export bundle. When no events match, the bundle is still valid:
            recordCount is 0, records is empty and the chain anchors are null.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditExportBundle'
        '400':
          description: Both filters supplied, or neither
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    AuditEventFields:
      type: object
      description: Fields shared between the audit event request and the persisted audit event.
      required:
        - eventType
        - actorId
        - resourceType
        - resourceId
      properties:
        eventType:
          type: string
          description: What happened (e.g., USER_LOGIN, RECORD_UPDATED, PERMISSION_GRANTED, USER_LOGOUT)
          examples:
            - USER_LOGIN
            - RECORD_UPDATED
            - PERMISSION_GRANTED
            - USER_LOGOUT
        actorId:
          type: string
          description: Who or what caused the event
          examples:
            - user-42
        resourceType:
          type: string
          description: The type of resource affected
          examples:
            - Customer
        resourceId:
          type: string
          description: The specific resource affected
          examples:
            - cust-1001
        payload:
          $ref: '#/components/schemas/EventPayload'
        timestamp:
          type: string
          format: date-time
          description: >-
            When the event occurred. May be supplied by the caller; if omitted
            in the request, the server assigns the time at which the event was
            recorded.
          examples:
            - '2026-09-14T10:15:30Z'

    EventPayload:
      type: object
      description: A structured object with event-specific detail.
      additionalProperties: true
      example:
        field: address
        oldValue: 123 Main St
        newValue: 456 Oak Ave

    AuditEventRequest:
      description: Audit event to record.
      allOf:
        - $ref: '#/components/schemas/AuditEventFields'

    AuditEventResponse:
      description: >-
        The persisted audit event, including server-assigned fields used for
        tamper-evident hash chaining.
      allOf:
        - $ref: '#/components/schemas/AuditEventFields'
        - type: object
          required:
            - id
            - sequenceNumber
            - timestamp
            - previousHash
            - contentHash
          properties:
            id:
              type: string
              format: uuid
              description: Server-assigned unique identifier of the audit event
              examples:
                - 3fa85f64-5717-4562-b3fc-2c963f66afa6
            sequenceNumber:
              type: integer
              format: int64
              description: Server-assigned monotonically increasing position of the event in the log
              examples:
                - 1024
            previousHash:
              type: string
              pattern: '^[a-f0-9]{64}$'
              description: SHA-256 content hash of the preceding event in the log (hex-encoded); links this event into the tamper-evident chain
              examples:
                - a94a8fe5ccb19ba61c4c0873d391e987982fbbd3a94a8fe5ccb19ba61c4c0873
            contentHash:
              type: string
              pattern: '^[a-f0-9]{64}$'
              description: SHA-256 hash (hex-encoded) of this event's canonical content, including previousHash
              examples:
                - 9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08

    AuditEventPage:
      type: object
      description: >-
        A single page of audit events together with the cursor needed to read the
        next page. No total count is reported: counting every matching row defeats
        the purpose of keyset pagination on an append-only log that only grows.
      required:
        - content
        - pageSize
        - hasMore
      properties:
        content:
          type: array
          description: The audit events on this page, in ascending sequence order
          items:
            $ref: '#/components/schemas/AuditEventResponse'
        pageSize:
          type: integer
          format: int32
          description: Maximum number of events per page used for this query
          examples:
            - 20
        nextCursor:
          type: integer
          format: int64
          description: >-
            Cursor to send as the cursor parameter to read the next page. It is
            the sequence number of the last event on this page, and is null when
            hasMore is false.
          examples:
            - 1024
        hasMore:
          type: boolean
          description: True when at least one further matching event exists after this page
          examples:
            - true

    AuditVerificationResponse:
      type: object
      description: Result of verifying the tamper-evident hash chain of the audit log.
      required:
        - valid
        - recordsChecked
      properties:
        valid:
          type: boolean
          description: True when the hash chain is intact and no tampering was detected
          examples:
            - true
        recordsChecked:
          type: integer
          format: int64
          description: >-
            Number of audit records inspected. When a violation is found,
            verification stops and this is the count checked up to and including
            the offending record.
          examples:
            - 15042
        firstViolation:
          description: Details of the first integrity violation found, or null when the log is valid.
          oneOf:
            - $ref: '#/components/schemas/HashChainViolation'
            - type: 'null'

    HashChainViolation:
      type: object
      description: The first point in the audit log at which integrity verification failed.
      required:
        - sequenceNumber
        - recordId
        - type
        - message
      properties:
        sequenceNumber:
          type: integer
          format: int64
          description: Sequence number of the offending audit record
          examples:
            - 42
        recordId:
          type: string
          description: Identifier of the offending audit record
          examples:
            - '019...'
        type:
          type: string
          description: Machine-readable kind of integrity violation
          examples:
            - CONTENT_HASH_MISMATCH
        message:
          type: string
          description: Human-readable explanation of the violation
          examples:
            - Stored content hash does not match canonical event content.

    AuditExportBundle:
      type: object
      description: >-
        A self-contained, independently verifiable export of the audit events
        matching one filter. The records carry the same fields the list API
        returns, including the hash-chain fields previousHash and contentHash,
        so a recipient can recompute every hash and check the chain anchors
        without access to this service or its database.
      required:
        - format
        - exportedAt
        - filter
        - chain
        - records
      properties:
        format:
          type: string
          description: Versioned identifier of the bundle format
          examples:
            - audit-bundle-v1
        exportedAt:
          type: string
          format: date-time
          description: When the bundle was produced
          examples:
            - '2026-09-15T08:30:00Z'
        filter:
          $ref: '#/components/schemas/AuditExportFilter'
        chain:
          $ref: '#/components/schemas/AuditExportChain'
        records:
          type: array
          description: The matching audit events, in ascending sequence order
          items:
            $ref: '#/components/schemas/AuditEventResponse'

    AuditExportFilter:
      type: object
      description: The single scope the export was filtered on.
      required:
        - type
        - value
      properties:
        type:
          type: string
          description: Which attribute the export was filtered on
          examples:
            - resourceId
            - actorId
        value:
          type: string
          description: The exact value every exported record matches
          examples:
            - cust-1001

    AuditExportChain:
      type: object
      description: >-
        Anchors pinning the exported records to the tamper-evident chain. For
        an empty export only algorithm and recordCount (0) are set; the anchor
        fields are null.
      required:
        - algorithm
        - recordCount
      properties:
        algorithm:
          type: string
          description: Hash algorithm used for contentHash and previousHash
          examples:
            - SHA-256
        firstRecordId:
          type: string
          format: uuid
          description: Identifier of the first exported record
        firstSequenceNumber:
          type: integer
          format: int64
          description: Sequence number of the first exported record
          examples:
            - 1024
        firstPreviousHash:
          type: string
          pattern: '^[a-f0-9]{64}$'
          description: >-
            previousHash of the first exported record, i.e. the content hash of
            the (possibly unexported) record that precedes it in the chain
        lastRecordId:
          type: string
          format: uuid
          description: Identifier of the last exported record
        lastSequenceNumber:
          type: integer
          format: int64
          description: Sequence number of the last exported record
          examples:
            - 1031
        lastRecordHash:
          type: string
          pattern: '^[a-f0-9]{64}$'
          description: contentHash of the last exported record
        recordCount:
          type: integer
          format: int64
          description: Number of records in the bundle
          examples:
            - 3

    ErrorResponse:
      type: object
      description: Error details returned when a request cannot be processed.
      required:
        - message
      properties:
        message:
          type: string
          description: Human-readable description of the error
          examples:
            - actorId is required
        code:
          type: string
          description: Machine-readable error code
          examples:
            - VALIDATION_FAILED